Update: migrated continers preset to dedicated file
