--- src/docker/containers/mysql.js
@@ -1,24 +1,0 @@
-const Container = require('./container') // eslint-disable-line no-unused-vars
-const { Docker } = require('../docker')
-
-/**
- * Create mysql continer.
- * @param {string} name - The name of the container.
- * @param {number} port - The port expose to the host.
- * @returns {Promise<Container>}
- */
-function CreateMysqlContainer (name, port, run = false) {
-  return Docker.prototype.CreateContainer({
-    image: 'mysql:5.7',
-    name: `cywp-${name}-mysql`,
-    network: 'cywp-network',
-    exposePorts: [
-      { host: port, docker: 3306 }
-    ],
-    environmentVariables: [
-      { name: 'MYSQL_ROOT_PASSWORD', value: 'cywp' }
-    ]
-  }, run)
-}
-
-module.exports = { CreateMysqlContainer }
